env_memdbg_alignment should be a size_t
Former-commit-id: b6df9c2bb2397f3cbdf84b6744c058574e9e3975

+++ include/Profile/TauEnv.h
@@ -100,7 +100,7 @@ extern "C" {
   size_t TAUDECL TauEnv_get_memdbg_alloc_max_value();
   int TAUDECL TauEnv_get_memdbg_overhead();
   size_t TAUDECL TauEnv_get_memdbg_overhead_value();
-  unsigned int TAUDECL TauEnv_get_memdbg_alignment();
+  size_t TAUDECL TauEnv_get_memdbg_alignment();
   int TAUDECL TauEnv_get_memdbg_zero_malloc();
 
 #ifdef __cplusplus
